Add QoirLib C++ DLL project
QoirLib.h is the original Qoir.h just with the qoir_encode and qoir_code functions modified so they are exported to a DLL. This is all then included in a project that can be used to compile the DLL using Visual Studio 2022 Community Edition. A DLL build (release config) is also included.