Fix potential vicious cycle on empty frame buffer
When the frame buffer ran dry a reset was executed, but the internal frame list wasn't fully cleared. This could lead to old audio frames still lingering in the buffer which would then forever not be returned to the output, completely breaking audio.

Also moved the Reset function close to the constructor so that it's immediately visible what is initialized where.