Changed "Receiver Filter" to "Sender Filter"
This is a breaking change. It will move existing filters from the "Audio/Video Filters" section down to the "Effect Filters" section, where they will appear in a broken state with no properties. You need to delete the broken filters and readd and reconfigure them in the "Audio/Video Filters" section.