Fix audio data plane and size calculations
WIth this unwritten rule that filters must be 10 times more complex than outputs, I should have anticipated that it's not enough for filters to have an ever changing frame-count with every packet, they also (unlike) outputs need to have their data in non-contiguous memory so that the plane pointers need to be considered. Say hello to another performance decrease also for outputs induced by the filter implementation.