Show warning and help for missing libjpeg-turbo
There was a bug that prevented the JPEG option from being hidden when the library was missing, but now instead of fixing this the option will stay visible but disabled, and a warning message and help button will be shown. This is more transparent for the user and hopefully confusions as in #28 can be avoided for the future.