Add warning for non-local network interfaces (1/2)
Show a warning on the settings dialog and in the log if either a network interface is selected that doesn't have a local address assigned (i.e. an Internet address) or "Any" was selected and none of the available interfaces have a local address.
Part 1 of 2 for senders (outputs and filters). Part 2 will be for receivers (sources).