Fix receive delay not calculated on zero latency
If the actual delay is zero the whole time (same PC) this bug wasn't causing any issue, the calculated delay would stay at its initial value of zero, which is correct.

However, if the latency temporarily goes up to e.g. 16 ms and then back to 0 it wouldn't be reset to zero but stay at 16. In theory this also affected render delay, but there a delay of 0 ms is unlikely anyway.