Fix receive/render delay calculation for high FPS
Under high FPS the receive timestamp packets could be too much to be processed fast enough and instead getting queued up in the receive buffer, making the delay calculation work on outdated timestamps from that queue and therefore showing too high delay values.

To solve this any excess timestamp data in the receive buffer is now discarded and only the latest data is read on every receive iteration.

This is a basic fail-safe that can stay in place and will always help regardless of system performance, configured FPS and any short hiccups. However, the timestamp packets in general should be looked at again, some optimization is in order, e.g. in general the frequency of updates should be lowered/capped and maybe also the receiving can be optimized (locking?).