Fix wrong render delay after receiver reconnect
If a source was hidden and the receiver disconnected, then reconnected after 5 seconds, the render delay would initially add 5 seconds to the calculcated delay for the first few frames without frame buffer, or for the whole time of frame buffer filling if a frame buffer was used.

This fixes it by resetting previous render timestamps on receiver connection and only remembering new render timestamps if data is currenty sent to the output.