Fix wrong pipe identifier for manual selections
When a pipe sender was selected from the feeds list in manual mode, this would preset the pipe identifier field with "identifer | type" instead of just "identifier", which is addressed by this fix.

Additionally the feeds list is now always reset to "No feed selected" after selecting an item in manual mode for both pipe and socket connections, as this selection is not meant to be persisted, in manual mode it only acts as a temporary helper to fill the other fields.