Fix client errors killing peer discovery server
UDP is connectionless, nevertheless in Windows when a UDP client resets it's "connection" this is signaled to the server using ICMP, causing a ConnectionReset SocketException there.

Since with UDP all socket errors are supposed to be non-recoverable the server is currently exiting its receive loop if that happens, meaning clients sending a connection reset can kill the peer discovery receive server.

This fix sets a socket option that makes the UDP socket ignore such ICMP packets so that the server is no longer affected by this.