Fix peer discovery requests with multiple NICs
When multiple functional network interfaces were available in a system, only one of them would be used to send peer discovery request messages. If that happened to be a virtual interface, no discovery messages would ever be sent from the real network card to the real network.

Now the code will loop through all network interfaces with multicast capability and send the discovery message from each of them. Duplicate responses will be ignored, in case multiple interfaces are on the same network and receive the same responses.

The same applies to listening for peer discovery requests, the server will now listen on all interfaces.